Crate re_format_arrow

Formatting for tables of Arrow arrays

Structs

DisplayMetadata 🔒

RecordBatchFormatOpts

Functions

custom_array_formatter 🔒

format_cell 🔒

format_dataframe_with_metadata 🔒

format_dataframe_without_metadata 🔒

format_record_batch

Nicely format this record batch in a way that fits the terminal.

format_record_batch_opts

Nicely format this record batch using the specified options.

format_record_batch_with_width

Nicely format this record batch, either with the given fixed width, or with the terminal width (None).

parse_tuid 🔒

trim_name 🔒

Type Aliases

CustomArrayFormatter 🔒

Format the given row as a string

Metadata 🔒

This is a BTreeMap, and not a HashMap, because we want a predictable order.