re_format_arrow
Function format_record_batch_with_width

pub fn format_record_batch_with_width(
    batch: &RecordBatch,
    width: Option<usize>,
) -> Table

Nicely format this record batch, either with the given fixed width, or with the terminal width (None).

If transposed is true, the dataframe will be printed transposed on its diagonal axis. This is very useful for wide (i.e. lots of columns), short (i.e. not many rows) datasets.